requirements.txt
phone==0.4.5
python-decouple==3.8
